explicit type conversion in public API
This silences warnings if compiled in C++

+++ src/core/eegdev-pluginapi.h
@@ -194,9 +194,9 @@ static inline
 void egdi_set_gval(union gval* dst, int type, double val)
 {
 	if (type == EGD_INT32)
-		dst->valint32_t = val;
+		dst->valint32_t = (int32_t) val;
 	else if (type == EGD_FLOAT)
-		dst->valfloat = val;
+		dst->valfloat = (float) val;
 	else if (type == EGD_DOUBLE)
 		dst->valdouble = val;
 }
